scripts: Add a west command to generate rpc file
We use protobuf-c to generate all the struct used to exchange
data with the host.
Unfortunatelly, protbuf-c doesn't generate any RPC and this should
be done manually. This adds a script, some templates and a west
command that could be used to generate the RPC.
This is going to be useful if we need to update them (ESPHome is evolving
a lot) or if need to update some part of the generated code.

This has many hacks or shortcuts but this works as is so far.
This will have to be improved at some poinnt.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>